Moved testimonial to its own folder. Styled the carousel.
