+++ src/containers/Home/Testimonials/Testimonials.css
@@ -1,0 +1,34 @@
+.carousel {
+  background-color: #506361;
+}
+
+.carousel-item {
+  height: 768px;
+}
+
+.carousel-caption {
+  top: 0;
+  margin: 0 auto;
+  max-width: 1310px;
+  padding: 100px 0 100px 0;
+}
+
+.carousel-caption-header {
+  font-size: 2.5rem;
+  font-weight: 700;
+  text-align: center;
+}
+
+.carousel-caption-body {
+  padding: 40px 0 0 0;
+  font-size: 1.5rem;
+}
+
+.carousel-caption-footer {
+  padding: 40px 0 0 0;
+  font-size: 1.5rem;
+}
+
+.testimonial-author {
+  font-weight: 700;
+}
